Add api version config option
