+++ ibm_mq/datadog_checks/ibm_mq/data/conf.yaml.example
@@ -85,6 +85,11 @@ instances:
     #
     # auto_discover_queues: false
 
+    ## param api_version - int - optional - default: 6
+    ## Which API version to use
+    #
+    # api_version: 6
+
     ## @param ssl_auth - string - optional
     ## Whether or not to use SSL auth while connecting to the channel.
     #
