+++ ibm_mq/tests/common.py
@@ -179,3 +179,13 @@ METRICS = (
 OPTIONAL_METRICS = [
     'ibm_mq.queue.max_channels',
 ]
+
+
+def assert_all_metrics(aggregator):
+    for metric, metric_type in METRICS:
+        aggregator.assert_metric(metric, metric_type=getattr(aggregator, metric_type.upper()))
+
+    for metric in OPTIONAL_METRICS:
+        aggregator.assert_metric(metric, at_least=0)
+
+    aggregator.assert_all_metrics_covered()
